requirements.txt
flask==2.3.3
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1 